dynamic key

+++ src/main/java/com/njust/fog_node/SaticScheduleTask.java
@@ -30,22 +30,23 @@ public class SaticScheduleTask {
 
     //添加定时任务
     @Scheduled(fixedDelay = 5000)
-    private void aggregatorTask() {
-        List<EncryptedDataForm> list = null;
+    private void aggregatorTask() throws Exception {
+        List<EncryptedDataForm> list;
         for(;;){
             list = edfDao.queryEncryptedDataForms();
+            if(list.size()==0) break;
             for(EncryptedDataForm encryptedDataForm : list){
-                System.out.println(encryptedDataForm.getId());
                 edfDao.deleteRawById(encryptedDataForm.getId());
                 encryptedDataForm.setId(null);
                 edfDao.addToUsed(encryptedDataForm);
             }
             EDataAggregator eda = new EDataAggregator();
-
-            PaillierPublicKey paillierPublicKey = PaillierPublicKey.paillierJsonToPublicKey("{\"n\":7037996759611275900405487329144489085210900622405788623915340046554895678557675360099993502545810105916795350348201798995744651664108236879690390748857833,\"nSquare\":49533398388298819693190911443085500113137594389227717398938303574532356291531019850234314622241175041250992063305927006862844026670633749958420794136365527887009273250901790502746504678689585917463571409706569379921923499464969602901871572009667889989146252127852333968575165007138552703354893437794045455889,\"g\":47,\"bitLength\":512,\"timeStamp\":1580452220178}");
+            //PaillierPublicKey paillierPublicKey = PaillierPublicKey.paillierJsonToPublicKey("{\"n\":7037996759611275900405487329144489085210900622405788623915340046554895678557675360099993502545810105916795350348201798995744651664108236879690390748857833,\"nSquare\":49533398388298819693190911443085500113137594389227717398938303574532356291531019850234314622241175041250992063305927006862844026670633749958420794136365527887009273250901790502746504678689585917463571409706569379921923499464969602901871572009667889989146252127852333968575165007138552703354893437794045455889,\"g\":47,\"bitLength\":512,\"timeStamp\":1580452220178}");
+            PaillierPublicKey paillierPublicKey = PaillierPublicKey.readFromFile();
             PaillierCalculator paillierCalculator = new PaillierCalculator(paillierPublicKey);
 
             rdDao.add(eda.eDataAggregator(list,paillierCalculator));
+
             if(list.size()<100){
                 System.out.println("执行静态定时聚合任务结束时间: " + LocalDateTime.now());
                 break;
@@ -62,12 +63,16 @@ public class SaticScheduleTask {
         RestTemplate restTemplate = new RestTemplate();
         String url = "http://localhost:8081/data/postRData";
         MultiValueMap<String, String> paramMap = new LinkedMultiValueMap<>();
-
         List<ResultData> list = rdDao.queryResultData();
+        if(list.size()==0) return;
         for(;;){
             for(ResultData resultData : list){
                 paramMap.add("data",gson.toJson(resultData));
-                System.out.println(restTemplate.postForObject(url,paramMap,String.class));
+                String response = restTemplate.postForObject(url,paramMap,String.class);
+                if(response.equals("WKTS")){
+                    System.out.println("密钥过期，重置密钥");
+                    PaillierPublicKey.renovate();
+                }
                 paramMap.clear();
                 rdDao.deleteRawById(resultData.getId());
             }
@@ -76,8 +81,6 @@ public class SaticScheduleTask {
                 break;
             }
         }
-
-
 
     }
